% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramky.R
\name{Ramky}
\alias{Ramky}
\title{This creates histogram,boxplot and numeric summary}
\usage{
Ramky(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
This creates histogram,boxplot and numeric summary
}
